Refine timestamp matchers. Update unit test to confirm expected match out of default detection list.
